requirements.txt
python-telegram-bot==20.1
selenium==4.9.0
webdriver-manager==4.8.1
